Change the way how location details are shown
TabbedPage (LocationPage) is removed and navigation from LocationListPage goes directly to LocationGraphPage. It has now ToolbarItem to navigate to LocationDetailsPage.

This change was mainly done as workaround to fix issue on iOS, where navigating from Shell page to TabbedPage does not work correctly due to an issue in XF.

On the other hand, the detailed location information is not that relevant anyway. The new approach is not perfect either, but at least it works on both platforms.