Copy dummy measurement data from last real measurement
This hack was needed to get layout shown correctly on iOS. Otherwise there's lines with background color around the dummy item.

Now the dummy item data is copied from last real data, so that the item has exactly same height. The text color for dummy item is set to same as surface.